// apps/oauth/src/infrastructure/Enums.ts
export enum TOKEN_TYPES {
  AUTHORIZATION_CODE = 'AUTHORIZATION_CODE',
  CONFIRMATION = 'CONFIRMATION',
  REFRESH = 'REFRESH',
  ACCESS = 'ACCESS',
  OAUTH = 'OAUTH',
  RESET = 'RESET',
}

export enum TEMPLATES {
  CONFIRM_EMAIL = 'confirmation',
}

export enum CONFIRM_EMAIL_STATUS {
  ALREADY_CONFIRMED = ' ALREADY_CONFIRMED',
  EXPIRED = 'EXPIRED',
  INVALID = 'INVALID',
  SUCCESS = 'SUCCESS',
}

export enum SIGN_UP_STATUS {
  FORBIDDEN = 'FORBIDDEN',
  SUCCESS = 'SUCCESS',
  ERROR = 'ERROR',
}

export enum STATUS {
  FAILED = 'FAILED',
  SUCCESS = 'SUCCESS',
}